RecordHub Service: move user code into submodule.
