+++ recordhub/customer/pom.xml
@@ -2,29 +2,22 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>org.springframework.boot</groupId>
-    <artifactId>spring-boot-starter-parent</artifactId>
-    <version>3.1.0</version>
-    <relativePath/> 
+    <groupId>com.mcalder.recordhub</groupId>
+    <artifactId>recordhub</artifactId>
+    <version>0.1</version>
   </parent>
-  <groupId>com.rez</groupId>
-  <artifactId>user</artifactId>
+
+  <artifactId>customer</artifactId>
   <version>0.2</version>
   <packaging>jar</packaging>
 
-  <name>user</name>
+  <name>customer</name>
   <properties>
 
     <!-- TODO Update to your own Docker repository or Docker Hub scope -->
     <dockerImage>registry.hub.docker.com/max8github/${project.artifactId}</dockerImage>
-    <dockerTag>${project.version}-${build.timestamp}</dockerTag>
-    <maven.build.timestamp.format>yyyyMMddHHmmss</maven.build.timestamp.format>
-    <mainClass>com.rez.user.Main</mainClass>
-
-    <jdk.target>17</jdk.target>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
-    <kalix-sdk.version>1.3.3</kalix-sdk.version>
+    <dockerTag>${project.version}-${maven.build.timestamp}</dockerTag>
+    <mainClass>com.mcalder.recordhub.customer.Main</mainClass>
   </properties>
 
   <build>
@@ -135,9 +128,36 @@
       </plugin>
 
       <plugin>
+        <groupId>org.codehaus.mojo</groupId>
+        <artifactId>exec-maven-plugin</artifactId>
+        <version>3.0.0</version>
+        <executions>
+          <execution>
+            <goals>
+              <goal>exec</goal>
+            </goals>
+          </execution>
+        </executions>
+        <configuration>
+          <executable>java</executable>
+          <arguments>
+            <argument>-classpath</argument>
+            <classpath/>
+            <argument>-Dlogging.config=src/main/resources/logback-dev-mode.xml</argument>
+            <argument>${mainClass}</argument>
+          </arguments>
+          <environmentVariables>
+            <!-- needed for the proxy to access the user function on all platforms -->
+            <HOST>0.0.0.0</HOST>
+          </environmentVariables>
+        </configuration>
+      </plugin>
+
+      <plugin>
         <!-- configure src/it/java and src/it/resources -->
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>build-helper-maven-plugin</artifactId>
+        <version>${build-helper-maven-plugin.version}</version>
         <executions>
           <execution>
             <id>add-integration-test-source</id>
@@ -171,17 +191,13 @@
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-surefire-plugin</artifactId>
+        <version>${maven-surefire-plugin.version}</version>
         <configuration>
           <excludes>
             <!-- ignore integration test classes -->
             <exclude>**/*IntegrationTest</exclude>
           </excludes>
         </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-maven-plugin</artifactId>
       </plugin>
 
       <plugin>
@@ -198,6 +214,23 @@
         <configuration>
           <dockerImage>${dockerImage}:${dockerTag}</dockerImage>
           <mainClass>${mainClass}</mainClass>
+          <integrationTestSourceDirectory>src/it/java</integrationTestSourceDirectory>
+        </configuration>
+      </plugin>
+      <plugin>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-maven-plugin</artifactId>
+        <configuration>
+          <systemPropertyVariables>
+            <logging.config>src/main/resources/logback-dev-mode.xml</logging.config>
+            <kalix.user-function-interface>0.0.0.0</kalix.user-function-interface>
+          </systemPropertyVariables>
+          <excludes>
+            <exclude>
+              <groupId>org.projectlombok</groupId>
+              <artifactId>lombok</artifactId>
+            </exclude>
+          </excludes>
         </configuration>
       </plugin>
       <plugin>
@@ -208,7 +241,6 @@
           <skip>true</skip>
         </configuration>
       </plugin>
-
     </plugins>
   </build>
 
@@ -254,17 +286,15 @@
       <version>${kalix-sdk.version}</version>
       <scope>test</scope>
     </dependency>
-
     <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter-api</artifactId>
+      <groupId>org.awaitility</groupId>
+      <artifactId>awaitility</artifactId>
+      <version>4.1.0</version>
       <scope>test</scope>
     </dependency>
     <dependency>
-      <groupId>org.junit.jupiter</groupId>
-      <artifactId>junit-jupiter</artifactId>
-      <scope>test</scope>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
     </dependency>
-    
   </dependencies>
 </project>
