+++ recordhub/customer/src/main/java/com/mcalder/recordhub/customer/user/domain/Address.java
@@ -1,3 +1,3 @@
-package com.rez.user.domain;
+package com.mcalder.recordhub.customer.user.domain;
 
 public record Address(String street, String city) {}
